packages/socket_forwarder/pyproject.toml

[project]
name = "radium226-socket_forwarder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-statemachine>=2.4.0",
]

[project.scripts]
forward-socket = "radium226.socket_forwarder:app"

[tool.uv]
package = true
dev-dependencies = [
  "psycopg2>=2.9.10",
  "pytest>=8.3.3",
  "python-statemachine>=2.4.0",
  "radium226-pg",
  "ruff>=0.8.1",
]
